requirements/requirements.txt
attrs>=22.0.0,<25.0.0
cattrs>=22.0.0,<25.0.0 # Same versioning as attrs
typeguard>=4.0.0,<5.0.0
pydantic>=2.7.1,<3.0.0
paramiko>=3.4.0,<4.0.0
pyyaml>=6.0.2,<7.0.0
more_itertools>=9.0.0,<11.0.0
adbutils>=2.0.0,<3.0.0
regex>=2021.11.10,<=2024.9.11 # Manually update